Use traits to organise the code
